requirements.txt
astor==0.8.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.6
cycler==0.10.0
decorator==5.1.0
gast==0.3.3
idna==3.2
kiwisolver==1.3.0
matplotlib==3.3.4
Naked==0.1.31
numpy==1.19.3
opencv-contrib-python==4.5.3.56
paddlepaddle==2.1.3
Pillow==8.3.2
protobuf==3.18.0
pycryptodome==3.10.4
PyEmail==0.0.1
pyparsing==2.4.7
python-dateutil==2.8.2
PyYAML==5.4.1
requests==2.26.0
schedule==1.1.0
shellescape==3.8.1
six==1.16.0
soupsieve==2.2.1
urllib3==1.26.7
